Make boolean ops support the Text type
